Fix logging_collector=on sed command on macOS.
According to https://stackoverflow.com/a/19457213 macOS's version of sed
expects that -i will take an argument, so the previous form was treating
the sed expression as the backup file extension. Specifying an empty
extension to disable backups (and for good measure making the expression
flag explicit) fixes the problem.