Switch back off lld
Per https://www.postgresql.org/message-id/20220223002638.vicup5bxeefaobg5%40alap3.anarazel.de
it seems to be the source of gdb slowdowns, so it's not worth using. The
(unmaintained) gold is still an option, but it's ELF-only and so doesn't
work on MacOS, so for now I'm punting on making this OS dependent.